completed wordle mechanics
There are still some bugs to be fixed and need to fix help modal and add modals for the settings and statistics. Also need to add statistics managing and need to add statistics to localstorage as well as a on-screen keyboard. Additionally, need to make the code more dynamic for language changing which will be in settings modal.